% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inks.R
\name{inks}
\alias{deltablur}
\alias{inks}
\alias{inkspark}
\title{Internal Function For Spatial KZ Filter \code{kzmd}}
\usage{
inks(ss, iterD, iterS, namex)

inkspark(ss, delta, sc)

deltablur(ss, delta)
}
\arguments{
\item{ss}{Data frame of signal values and positions.}

\item{iterD}{Vector. Window size for KZ filter.}

\item{iterS}{Vector. Scales of each dimension.}

\item{namex}{Names for each column of input data.}

\item{delta}{Integer. Window size of current dimension.}

\item{sc}{Integer. Scales of current dimension.}
}
\description{
\code{inks} is the control structure of the \code{kzmd} program.
It returns data frame of signal values and their grid positions.

\code{inkspark} prepares the expanded data frame for KZ filtering.

\code{deltablur} calculates the filtered value of each grid point.
}
\keyword{internal}

