% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kzfs.R
\docType{data}
\name{kzp2.QF}
\alias{kzp2.QF}
\alias{kzp2.demo}
\alias{kzp2.QF}
\title{The Demo Dataset For Examples of \code{optD2R}}
\format{An object of class \code{matrix} with 150 rows and 150 columns.}
\usage{
kzp2.demo

kzp2.QF
}
\description{
Datasets containing the original and optimized 2D periodogram spike records 
output by function \code{kzp2} and \code{optD2R},respectively. 
They are only used for saving running-time
of the examples. The following code is used to generate this dataset:
\itemize{
\item \code{kzp2.demo <- kzp2(a)$kzp2d}
\item \code{kzp2.QF <- optD2R(a, kzp2.summary(kzp2.demo, num=2))}
}
where \code{a} is the spatial data array for a wave field.
}
\seealso{
\code{kzpdr.demo}, \code{kzpdr.QF}
}
\keyword{datasets}
