% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kzfs.R
\docType{data}
\name{kzpdr.QF}
\alias{kzpdr.QF}
\alias{kzpdr.demo}
\title{The Demo Dataset For Examples of \code{kzpdr} and \code{optDR}}
\format{A list for a data frame and its MD5sum value.
The data frame \code{rec} has 18 rows and a few variables:
\describe{
  \item{direction}{angle of sampling, in degree}
  \item{freq}{frequency values of spikes}
  \item{spg}{smoothed power periodogram values of spikes}
  \item{...}{...}
}}
\usage{
kzpdr.demo

kzpdr.QF
}
\description{
Datasets containing the spectral spike records of directional periodograms
output by function \code{kzpdr} and \code{optDR}, respectively. 
They are only used for saving running-time
of the examples. The following code is used to generate this dataset:
\itemize{
\item \code{kzpdr.demo <- kzpdr(a, c(0, pi/4, pi/3, -pi/3, pi/18), plot=TRUE)}
\item \code{kzpdr.QF <- optDR(a, kzpdr.demo)}
}
where \code{a} is the spatial data array for a wave field.
}
\seealso{
\code{kzp2.demo}, \code{kzp2.QF}
}
\keyword{datasets}
